+++ Demo/Demo/DemoScreen.swift
@@ -3,16 +3,14 @@
 //  ScrollKit
 //
 //  Created by Daniel Saidi on 2023-02-04.
-//  Copyright © 2023-2024 Daniel Saidi. All rights reserved.
+//  Copyright © 2023-2025 Daniel Saidi. All rights reserved.
 //
 
 import ScrollKit
 import SwiftUI
 
-/**
- This view takes a custom header view and height and adds it
- to a scroll view with sticky header.
- */
+/// This view takes a custom header view and height and adds
+/// it to a scroll view with sticky header.
 struct DemoScreen<HeaderView: View>: View {
 
     let headerHeight: CGFloat
@@ -21,7 +19,7 @@ struct DemoScreen<HeaderView: View>: View {
     let headerView: () -> HeaderView
 
     @State
-    private var headerVisibleRatio: CGFloat = 1
+    private var visibleHeaderRatio: CGFloat = 1
 
     @State
     private var scrollOffset: CGPoint = .zero
@@ -32,17 +30,27 @@ struct DemoScreen<HeaderView: View>: View {
         ScrollViewWithStickyHeader(
             header: header,
             headerHeight: headerHeight,
+            headerMinHeight: 75,
             scrollManager: scrollManager,
             onScroll: handleScrollOffset
         ) {
-            listItems
+            LazyVStack(spacing: 0) {
+                ForEach(1...100, id: \.self) { item in
+                    VStack(spacing: 0) {
+                        Text("Item \(item)")
+                            .padding()
+                            .frame(maxWidth: .infinity, alignment: .leading)
+                        Divider()
+                    }
+                }
+            }
         }
         .toolbar {
             ToolbarItem(placement: .principal) {
                 Text("Demo Title")
                     .font(.headline)
                     .previewHeaderContent()
-                    .opacity(1 - headerVisibleRatio)
+                    .opacity(1 - visibleHeaderRatio)
             }
             ToolbarItem(placement: .topBarTrailing) {
                 Button {
@@ -82,25 +90,12 @@ struct DemoScreen<HeaderView: View>: View {
             Text("Some additional information")
         }
         .padding(20)
-        .opacity(headerVisibleRatio)
+        .opacity(visibleHeaderRatio)
     }
 
-    var listItems: some View {
-        LazyVStack(spacing: 0) {
-            ForEach(1...100, id: \.self) { item in
-                VStack(spacing: 0) {
-                    Text("Item \(item)")
-                        .padding()
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                    Divider()
-                }
-            }
-        }
-    }
-
-    func handleScrollOffset(_ offset: CGPoint, headerVisibleRatio: CGFloat) {
+    func handleScrollOffset(_ offset: CGPoint, visibleHeaderRatio: CGFloat) {
         self.scrollOffset = offset
-        self.headerVisibleRatio = headerVisibleRatio
+        self.visibleHeaderRatio = visibleHeaderRatio
     }
 }