+++ Sources/ScrollKit/Examples/Examples+SpotifyAlbumScreen.swift
@@ -1,42 +1,53 @@
 //
-//  Spotify+PreviewScreen.swift
+//  Examples+SpotifyAlbumScreen.swift
 //  ScrollKit
 //
 //  Created by Daniel Saidi on 2023-02-07.
-//  Copyright © 2023-2024 Daniel Saidi. All rights reserved.
+//  Copyright © 2023-2025 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI
 
-public extension Spotify {
+public extension Examples.Spotify {
     
     /// This view mimics a Spotify album screen.
-    struct PreviewScreen: View {
+    struct AlbumScreen: View {
         
-        public init(info: PreviewInfo) {
-            self.info = info
+        public init(album: Album) {
+            self.album = album
         }
         
-        private var info: PreviewInfo
+        private var album: Album
+        
+        @Environment(\.colorScheme)
+        private var colorScheme
         
         @Environment(\.dismiss)
         private var dismiss
         
         @State
-        private var headerVisibleRatio = 1.0
+        private var visibleHeaderRatio = 1.0
         
         @State
         private var scrollOffset = CGPoint.zero
         
+        private var scrollContentCornerRadius: Double {
+            colorScheme == .dark ? 0.0 : 20
+        }
+        
         public var body: some View {
             ScrollViewWithStickyHeader(
                 header: scrollViewHeader,
-                headerHeight: Spotify.PreviewScreenHeader.height,
+                headerHeight: Examples.Spotify.AlbumScreen.Header.height,
+                headerMinHeight: 50,
+                headerStretch: false,
+                contentCornerRadius: scrollContentCornerRadius,
                 onScroll: handleScrollOffset
             ) {
-                Spotify.PreviewScreenContent(info: info)
+                if #available(iOS 16.0, *) {
+                    Examples.Spotify.AlbumScreen.Content(album: album)
+                }
             }
-            .preferredColorScheme(.dark)
             #if os(iOS)
             .hideBackButtonText()
             #endif
@@ -46,21 +57,22 @@ public extension Spotify {
         }
         
         func scrollViewHeader() -> some View {
-            Spotify.PreviewScreenHeader(
-                info: info,
-                headerVisibleRatio: headerVisibleRatio
+            Examples.Spotify.AlbumScreen.Header(
+                album: album,
+                bottomPadding: scrollContentCornerRadius,
+                visibleHeaderRatio: visibleHeaderRatio
             )
         }
         
         var toolbarTitleView: some View {
-            Text(info.releaseTitle)
+            Text(album.releaseTitle)
                 .font(.headline.bold())
-                .opacity(headerVisibleRatio > 0 ? 0 : -5 * headerVisibleRatio)
+                .opacity(visibleHeaderRatio > 0 ? 0 : -5 * visibleHeaderRatio)
         }
         
-        func handleScrollOffset(_ offset: CGPoint, headerVisibleRatio: CGFloat) {
+        func handleScrollOffset(_ offset: CGPoint, visibleHeaderRatio: CGFloat) {
             self.scrollOffset = offset
-            self.headerVisibleRatio = headerVisibleRatio
+            self.visibleHeaderRatio = visibleHeaderRatio
         }
     }
 }
@@ -90,14 +102,31 @@ private extension View {
     }
 }
 
-#Preview("Navigation") {
+private struct Preview: View {
+    
+    var body: some View {
+        Examples.Spotify.AlbumScreen(album: .misfortune)
+    }
+}
+
+#Preview("Light") {
+
+    Preview()
+}
+
+#Preview("Dark") {
+
+    Preview().preferredColorScheme(.dark)
+}
+
+#Preview("Push") {
 
     NavigationView {
         #if os(macOS)
         Color.clear
         #endif
         NavigationLink("Test") {
-            Spotify.PreviewScreen(info: .regina)
+            Preview()
         }
     }
     #if os(iOS)
@@ -108,7 +137,7 @@ private extension View {
 #if os(iOS)
 #Preview("Sheet") {
     
-    struct Preview: View {
+    struct SheetPreview: View {
         
         @State var isPresented = false
         
@@ -118,14 +147,14 @@ private extension View {
             }
             .sheet(isPresented: $isPresented) {
                 NavigationView {
-                    Spotify.PreviewScreen(info: .regina)
+                    Preview()
                 }
                 .navigationViewStyle(.stack)
             }
         }
     }
     
-    return Preview()
+    return SheetPreview()
 }
 #endif